requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
dnspython==2.7.0
eventlet==0.40.0
exceptiongroup==1.3.0
filelock==3.13.1
fsspec==2024.6.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
kombu==5.5.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
pillow==11.0.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
redis==6.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.3
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
